chore: Update file for transferring images, videos, and audios
Update the `gallerieViews.vue` file to allow transferring not only images but also videos and audios. This improves the functionality of the file by expanding the types of files that can be transferred. The changes align with recent user commits and repository commits that updated the navigation menu styles and colors.